+++ src/main/kotlin/com/jfrog/conan/plugin/models/Component.kt
@@ -1,0 +1,7 @@
+package com.jfrog.conan.plugin.models
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class Component(
+        val cmake_target_name: String? = null
+)